requirements.txt
Flask==2.2.2 #Use this version with Werkzeug 2.3.7 for Python3.9 on Azure Web App
Flask-Cors==3.0.10
flask-restful==0.3.8
Werkzeug==2.3.7 #Required to fix Werkzegu dependency issue on Azure Web App
flask_jwt_extended==4.3.1
flask-restful==0.3.8
requests==2.24.0
openai== 0.27.4
azure-storage-blob==12.9.0
azure-data-tables==12.4.1
azure-identity==1.7.0
python-dotenv==0.19.2